Replace "terraform state list" with "terraform show"
An orphaned resource file can exist for 2 different reasons:
1. Resource has been removed out-of-band from Infrakit, in this case
the resource in the terraform state file will be removed during a
`terraform refresh`.
2. Infrakit failed over to a different manager during a `terraform apply`,
in this case the state file is not updated because the apply never finished.

In order to differentiate between these 2 cases we need to get a list of
existing resources AND the associated IDs before we execute `terraform refresh`.

This commit simply replaces the command and the associated processing.

Signed-off-by: Steven Kaufer <kaufer@us.ibm.com>