Query backend for VM when ID is missing
Terraform does not write out the state file until after the provision is
complete. To get around this, we can query the backend for VMs based on the
unique tags.

This commit adds a map that will track the VM resource names to the ID, once
the backend ID is queried the map will cache it.

Signed-off-by: Steven Kaufer <kaufer@us.ibm.com>